Fix issue in the date picker is selectable lambda
[DESCRIPTION]
This issue happended because we were not considering that
the user can change the system date and time in the device
config. So, we will consider a date selectable if it is
at least bigger than the system date or the oldest added
dialysis AND less than the maximum dialysis or the current
system date.